+++ lib/ectomancer.ex
@@ -177,6 +177,7 @@ defmodule Ectomancer do
       @before_compile Ectomancer
 
       import Ectomancer.Tool, only: [tool: 2, authorize: 1]
+      import Ectomancer.Resource, only: [resource: 2]
       import Ectomancer.Expose, only: [expose: 1, expose: 2]
       import Ectomancer.RouteIntrospection, only: [expose_routes: 1, expose_routes: 2]
       import Ectomancer.ObanBridge, only: [expose_oban_jobs: 0, expose_oban_jobs: 1]